Update Caliban's module-running syntax
This PR replaces the old syntax of

caliban run -m <module_name> -p <containing_package>

with a required argument, module. We do some work now inside Caliban to resolve the argument into an actual module name to run and a containing package that we'll need to copy in to the docker container on builds.

Change-Id: I5927adcc81ce7da24157869c1b14049dcc908877