Add gsutil to our base docker images for Caliban
This PR gets gsutil installed. I was hesitant about this at first,
but gsutil can do things the python libraries can't, and the pytorch
official gcloud examples all use it. Docker lets us ship self-contained
environments... and it's a reasonable dependency to insist on for
others to install if they want to operate outside of docker.

Change-Id: Ibf0e2f04aa0416df22dd796bc10f6d544d460ecd