Add IPC Host option
this PR adds the `--ipc=host` option to all local docker tasks.

Without this jobs are crashing due to a lack of memory. With it they
get access to all memory in the host machine.

Benham turned me on to this... there are lots of references to folks needing this for
model training. Here's one example: https://gitlab.com/gitlab-org/gitlab-runner/issues/2168

I think this is a sane default that everyone should be using. then local commands
will match the user's expectation.

Change-Id: I19f67a444601c1aea16e089a648949483df4275d