[refs #00000] Allow editable installation of packages
Because pip does not recognize conda environments as proper virtual
environments, prior to this fix whenever it tried to install a package
in editable mode (with the -e flag), it would put the package in the
current working directory.  This caused problems because that directory
was overwritten when using caliban shell.

We fix this by setting the environmental variable PIP_SRC to be
/opt/conda/envs/caliban/pipsrc.  This tells pip to install the package
in that directory.