added non-none project check to caliban launcher (#88)
This PR adds a non-none check to the caliban launcher script. This fixes a common error case where a user creates a google.cloud.storage.Client() instance, which fails if no gcp project can be determined automatically. We thus add a check in the launcher script to set a project via the GOOGLE_CLOUD_PROJECT environment variable only in the case where no default project can be determined.