modifies caliban stop command to check with user up-front, rather than after querying running jobs (#90)
This PR modifies the caliban stop command to check with the user up-front rather than after querying running jobs, as this can take a long time, particularly with CAIP jobs. The modified behavior is to verify immediately, then stop all running and submitted jobs in the requested xgroup or user group.